+++ src/borneo/config.py
@@ -15,8 +15,10 @@ from time import sleep, time
 from typing import Callable
 
 try:
+    # noinspection PyCompatibility
     from urlparse import urlparse
 except ImportError:
+    # noinspection PyUnresolvedReferences,PyCompatibility
     from urllib.parse import urlparse
 
 from .auth import AuthorizationProvider
@@ -589,6 +591,7 @@ class StatsProfile(Enum):
     ALL = 4
 
 
+# noinspection PyPep8
 class NoSQLHandleConfig(object):
     """
     An instance of this class is required by :py:class:`NoSQLHandle`.
@@ -623,10 +626,10 @@ class NoSQLHandleConfig(object):
        endpoint argument)
      * nosql.us-ashburn-1.oci.oraclecloud.com (equivalent to using Region
        Regions.US_ASHBURN_1 as the endpoint argument)
-     * https\://nosql.us-ashburn-1.oci.oraclecloud.com:443
+     * https:\//nosql.us-ashburn-1.oci.oraclecloud.com:443
      * localhost:8080 - used for connecting to a Cloud Simulator instance
        running locally on port 8080
-     * https\://machine-hosting-proxy:443
+     * https:\//machine-hosting-proxy:443
 
     When using the endpoint (vs region id) syntax, if the port is omitted, the
     endpoint uses 8080 if protocol is http, and 443 in all other cases. If the
@@ -667,6 +670,7 @@ class NoSQLHandleConfig(object):
 
     def __init__(self, endpoint=None, provider=None):
         # Inits a NoSQLHandleConfig object.
+        endpoint_str = None
         if endpoint is not None:
             if not isinstance(endpoint, (str, Region)):
                 raise IllegalArgumentException(
@@ -679,6 +683,7 @@ class NoSQLHandleConfig(object):
                 self._region = Regions.from_region_id(endpoint)
             else:
                 self._region = endpoint
+                endpoint_str = endpoint.get_region_id()
             if self._region is None:
                 ep = endpoint
             else:
@@ -687,7 +692,8 @@ class NoSQLHandleConfig(object):
                     if (region_in_provider is not None and
                             region_in_provider != self._region):
                         raise IllegalArgumentException(
-                            'Specified region ' + endpoint + ' doesn\'t ' +
+                            'Specified region, ' + endpoint_str +
+                            ', doesn\'t ' +
                             'match the region in SignatureProvider.')
                 ep = self._region.endpoint()
         elif provider is not None: